updated pagination logics

+++ backend/internal/core/repository/paginated-response.go
@@ -1,9 +1,9 @@
 package repository
 
 type Pagination struct {
-	TotalCount int `json:"total_count"`
-	Page       int `json:"page"`
-	Limit      int `json:"limit"`
+	TotalCount int64 `json:"total_count"`
+	Page       int   `json:"page"`
+	Limit      int   `json:"limit"`
 }
 
 type PaginatedEntities[T any] struct {
